+++ apps/koku-ui-hccm/src/routes/optimizations/efficiency/efficiency.tsx
@@ -1,4 +1,5 @@
 import { Alert, AlertActionCloseButton, AlertActionLink, Grid, GridItem, PageSection } from '@patternfly/react-core';
+import AsyncComponent from '@redhat-cloud-services/frontend-components/AsyncComponent';
 import type { Providers } from 'api/providers';
 import { ProviderType } from 'api/providers';
 import type { OcpQuery } from 'api/queries/ocpQuery';
@@ -266,6 +267,7 @@ const Efficiency: React.FC<EfficiencyProps> = () => {
             variant="info"
           />
         )}
+        <AsyncComponent scope="costManagementRos" module="./FleetSummaryCards" />
         <Grid hasGutter>
           <GridItem xl={gridColumns}>
             <ComputeCard